backend/requirements.txt
pandas 
numpy 
yfinance
python-dotenv 
sqlalchemy 
psycopg2-binary
passlib[bcrypt]
fastapi==0.101.0
uvicorn==0.22.0
pydantic<2
fastapi-jwt-auth==0.5.0
bcrypt==4.0.1